+++ R/ggplot_sf_fun.R
@@ -208,14 +208,13 @@ ggplot_sf <- function(data,
 #' @param data A sf object with defined coordinate reference system. Required input.
 #' @param col_var Unquoted variable for points to be coloured by. Required input.
 #' @param col_method The method of colouring features, either "bin", "quantile" or "category." NULL results in "category", if categorical or "quantile" if numeric col_var. Note all numeric variables are cut to be inclusive of the min in the range, and exclusive of the max in the range (except for the final bucket which includes the highest value).
-#' @param bin_cuts A vector of bin cuts applicable where col_method of "bin" is selected. The first number in the vector should be either -Inf or 0, and the final number Inf. If NULL, 'pretty' breaks are used.
-#' @param quantile_cuts A vector of probability cuts applicable where col_method of "quantile" is selected. The first number in the vector should 0 and the final number 1. Defaults to quartiles. Only applicable where col_method equals "quantile".
-#' @param size Size of points. Defaults to 0.5.
-#' @param alpha The opacity of polygons. Defaults to 0.9.
-#' @param pal Character vector of hex codes. Defaults to NULL, which selects the colorbrewer Set1 or viridis.
-#' @param pal_rev Reverses the palette. Defaults to FALSE.
+#' @param col_cuts A vector of cuts to colour a numeric variable. If "bin" is selected, the first number in the vector should be either -Inf or 0, and the final number Inf. If "quantile" is selected, the first number in the vector should be 0 and the final number should be 1. Defaults to quartiles. 
 #' @param col_drop TRUE or FALSE  of whether to drop unused levels from the legend. Defaults to FALSE.
 #' @param col_na_remove TRUE or FALSE  of whether to remove NAs of the colour variable. Defaults to FALSE.
+#' @param pal Character vector of hex codes. Defaults to NULL, which selects the colorbrewer Set1 or viridis.
+#' @param pal_rev Reverses the palette. Defaults to FALSE.
+#' @param size Size of points. Defaults to 0.5.
+#' @param alpha The opacity of polygons. Defaults to 0.9.
 #' @param coastline Add a sf object as a coastline (or administrative boundaries). Defaults to NULL. Use nz (or nz_region) to add a new zealand coastline. Or add a custom sf object.
 #' @param coastline_behind TRUE or FALSE  as to whether the coastline is to be behind the sf object defined in the data argument. Defaults to FALSE.
 #' @param coastline_pal Colour of the coastline. Defaults to "#7F7F7F".
@@ -238,11 +237,11 @@ ggplot_sf <- function(data,
 #' @export
 #' @examples
 #' ggplot_sf_col(data = example_sf_nz_livestock, col_var = dairydens, coastline = nz,
-#'      col_method = "bin", bin_cuts = c(0, 10, 50, 100, 150, 200, Inf), legend_digits = 0,
+#'      col_method = "bin", col_cuts = c(0, 10, 50, 100, 150, 200, Inf), legend_digits = 0,
 #'      title = "Dairy density in count per km\u00b2, 2017")
 #'
 #' ggplot_sf_col(data = example_sf_nz_livestock, col_var = dairydens, coastline = nz,
-#'      col_method = "quantile", quantile_cuts = c(0, 0.25, 0.5, 0.75, 0.95, 1),
+#'      col_method = "quantile", col_cuts = c(0, 0.25, 0.5, 0.75, 0.95, 1),
 #'      title = "Dairy density in count per km\u00b2, 2017")
 #'
 #' map_data <- example_sf_nz_river_wq %>%
@@ -256,14 +255,13 @@ ggplot_sf <- function(data,
 ggplot_sf_col <- function(data,
                           col_var,
                           col_method = NULL,
-                          bin_cuts = NULL,
-                          quantile_cuts = c(0, 0.25, 0.5, 0.75, 1),
-                          size = 0.5,
-                          alpha = 0.9,
-                          pal = NULL,
-                          pal_rev = FALSE,
+                          col_cuts = NULL,
                           col_drop = FALSE,
                           col_na_remove = FALSE,
+                          pal = NULL,
+                          pal_rev = FALSE,
+                          size = 0.5,
+                          alpha = 0.9,
                           coastline = NULL,
                           coastline_behind = TRUE,
                           coastline_pal = "#7f7f7f",
@@ -337,40 +335,43 @@ ggplot_sf_col <- function(data,
     if (is.null(legend_labels)) labels <- waiver()
   }
   else if (col_method == "bin") {
-    if (!is.null(bin_cuts)) {
-      if (!(dplyr::first(bin_cuts) %in% c(0,-Inf))) warning("The first element of the bin_cuts vector should generally be 0 (or -Inf if there are negative values)")
-      if (dplyr::last(bin_cuts) != Inf) warning("The last element of the bin_cuts vector should generally be Inf")
+    if (!is.null(col_cuts)) {
+      if (!(dplyr::first(col_cuts) %in% c(0,-Inf))) warning("The first element of the col_cuts vector should generally be 0 (or -Inf if there are negative values)")
+      if (dplyr::last(col_cuts) != Inf) warning("The last element of the col_cuts vector should generally be Inf")
     }
-    if (is.null(bin_cuts)) bin_cuts <- pretty(col_var_vector)
+    if (is.null(col_cuts)) col_cuts <- pretty(col_var_vector)
     
     data <- dplyr::mutate(data,
                     !!col_var := cut(
                       col_var_vector,
-                      bin_cuts,
+                      col_cuts,
                       right = FALSE,
                       include.lowest = TRUE
                     ))
     
-    if (is.null(pal)) pal <- viridis::viridis(length(bin_cuts) - 1)
-    if (is.null(legend_labels)) labels <- numeric_legend_labels(bin_cuts, legend_digits)
+    if (is.null(pal)) pal <- viridis::viridis(length(col_cuts) - 1)
+    if (is.null(legend_labels)) labels <- numeric_legend_labels(col_cuts, legend_digits)
     if (!is.null(legend_labels)) labels <- legend_labels
   }
   else if (col_method == "quantile") {
-    if (dplyr::first(quantile_cuts) != 0) warning("The first element of the quantile_cuts vector generally always be 0")
-    if (dplyr::last(quantile_cuts) != 1) warning("The last element of the quantile_cuts vector should generally be 1")
-    bin_cuts <- quantile(col_var_vector, probs = quantile_cuts, na.rm = TRUE)
-    if (anyDuplicated(bin_cuts) > 0) stop("quantile_cuts do not provide unique breaks")
+    if(is.null(col_cuts)) col_cuts <- seq(0, 1, 0.25)
+    else {
+      if (dplyr::first(col_cuts) != 0) warning("The first element of the col_cuts vector generally always be 0")
+      if (dplyr::last(col_cuts) != 1) warning("The last element of the col_cuts vector should generally be 1")
+    }  
+    col_cuts <- quantile(col_var_vector, probs = col_cuts, na.rm = TRUE)
+    if (anyDuplicated(col_cuts) > 0) stop("col_cuts do not provide unique breaks")
     
     data <- dplyr::mutate(data,
                     !!col_var := cut(
                       col_var_vector,
-                      bin_cuts,
+                      col_cuts,
                       right = FALSE,
                       include.lowest = TRUE
                     ))
     
-    if (is.null(pal)) pal <- viridis::viridis(length(bin_cuts) - 1)
-    if (is.null(legend_labels)) labels <- numeric_legend_labels(bin_cuts, legend_digits)
+    if (is.null(pal)) pal <- viridis::viridis(length(col_cuts) - 1)
+    if (is.null(legend_labels)) labels <- numeric_legend_labels(col_cuts, legend_digits)
     if (!is.null(legend_labels)) labels <- legend_labels
   }
   
@@ -616,15 +617,14 @@ ggplot_sf_facet <- function(data,
 #' @param col_var Unquoted variable for points to be coloured by. Required input.
 #' @param facet_var Unquoted categorical variable to facet the data by. Required input.
 #' @param col_method The method of colouring features, either "bin", "quantile" or "category." NULL results in "category", if categorical or "quantile" if numeric col_var. Note all numeric variables are cut to be inclusive of the min in the range, and exclusive of the max in the range (except for the final bucket which includes the highest value).
-#' @param bin_cuts A vector of bin cuts applicable where col_method of "bin" is selected. The first number in the vector should be either -Inf or 0, and the final number Inf. If NULL, 'pretty' breaks are used.
-#' @param quantile_cuts A vector of probability cuts applicable where col_method of "quantile" is selected. The first number in the vector should 0 and the final number 1. Defaults to quartiles. Only applicable where col_method equals "quantile".
-#' @param quantile_by_facet TRUE of FALSE  whether quantiles should be calculated for each group of the facet variable. Defaults to TRUE.
-#' @param size Size of points. Defaults to 0.5.
-#' @param alpha The opacity of polygons. Defaults to 0.9.
-#' @param pal Character vector of hex codes. Defaults to NULL, which selects the colorbrewer Set1 or viridis.
-#' @param pal_rev Reverses the palette. Defaults to FALSE.
+#' @param col_cuts A vector of cuts to colour a numeric variable. If "bin" is selected, the first number in the vector should be either -Inf or 0, and the final number Inf. If "quantile" is selected, the first number in the vector should be 0 and the final number should be 1. Defaults to quartiles. 
+#' @param col_quantile_by_facet TRUE of FALSE  whether quantiles should be calculated for each group of the facet variable. Defaults to TRUE.
 #' @param col_drop TRUE or FALSE  of whether to drop unused levels from the legend. Defaults to FALSE.
 #' @param col_na_remove TRUE or FALSE  of whether to remove NAs of the colour variable. Defaults to FALSE.
+#' @param pal Character vector of hex codes. Defaults to NULL, which selects the colorbrewer Set1 or viridis.
+#' @param pal_rev Reverses the palette. Defaults to FALSE.
+#' @param size Size of points. Defaults to 0.5.
+#' @param alpha The opacity of polygons. Defaults to 0.9.
 #' @param coastline Add a sf object as a coastline (or administrative boundaries). Defaults to NULL. Use nz (or nz_region) to add a new zealand coastline. Or add a custom sf object.
 #' @param coastline_behind TRUE or FALSE  as to whether the coastline is to be behind the sf object defined in the data argument. Defaults to FALSE.
 #' @param coastline_pal Colour of the coastline. Defaults to "#7F7F7F".
@@ -660,15 +660,14 @@ ggplot_sf_col_facet <- function(data,
                                 col_var,
                                 facet_var,
                                 col_method = NULL,
-                                bin_cuts = NULL,
-                                quantile_cuts = c(0, 0.25, 0.5, 0.75, 1),
-                                quantile_by_facet = TRUE,
-                                size = 0.5,
-                                alpha = 0.9,
-                                pal = NULL,
-                                pal_rev = FALSE,
+                                col_cuts = NULL,
+                                col_quantile_by_facet = TRUE,
                                 col_drop = FALSE,
                                 col_na_remove = FALSE,
+                                pal = NULL,
+                                pal_rev = FALSE,
+                                size = 0.5,
+                                alpha = 0.9,
                                 facet_nrow = NULL,
                                 legend_ncol = 3,
                                 legend_digits = 1,
@@ -746,63 +745,66 @@ ggplot_sf_col_facet <- function(data,
     if (is.null(legend_labels)) labels <- waiver()
   }
   else if (col_method == "bin") {
-    if (!is.null(bin_cuts)) {
-      if (!(dplyr::first(bin_cuts) %in% c(0,-Inf))) warning("The first element of the bin_cuts vector should generally be 0 (or -Inf if there are negative values)")
-      if (dplyr::last(bin_cuts) != Inf) warning("The last element of the bin_cuts vector should generally be Inf")
+    if (!is.null(col_cuts)) {
+      if (!(dplyr::first(col_cuts) %in% c(0,-Inf))) warning("The first element of the col_cuts vector should generally be 0 (or -Inf if there are negative values)")
+      if (dplyr::last(col_cuts) != Inf) warning("The last element of the col_cuts vector should generally be Inf")
     }
-    if (is.null(bin_cuts)) bin_cuts <- pretty(col_var_vector)
+    if (is.null(col_cuts)) col_cuts <- pretty(col_var_vector)
     
       data <- dplyr::mutate(data,
                             !!col_var := cut(
                               col_var_vector,
-                              bin_cuts,
+                              col_cuts,
                               right = FALSE,
                               include.lowest = TRUE
                             ))
 
-    if (is.null(pal)) pal <- viridis::viridis(length(bin_cuts) - 1)
-    if (is.null(legend_labels)) labels <- numeric_legend_labels(bin_cuts, legend_digits)
+    if (is.null(pal)) pal <- viridis::viridis(length(col_cuts) - 1)
+    if (is.null(legend_labels)) labels <- numeric_legend_labels(col_cuts, legend_digits)
     if (!is.null(legend_labels)) labels <- legend_labels
   }
   else if (col_method == "quantile") {
-    if (dplyr::first(quantile_cuts) != 0) warning("The first element of the quantile_cuts vector generally always be 0")
-    if (dplyr::last(quantile_cuts) != 1) warning("The last element of the quantile_cuts vector should generally be 1")
-    if (quantile_by_facet == TRUE) {
+    if(is.null(col_cuts)) col_cuts <- seq(0, 1, 0.25)
+    else {
+      if (dplyr::first(col_cuts) != 0) warning("The first element of the col_cuts vector generally always be 0")
+      if (dplyr::last(col_cuts) != 1) warning("The last element of the col_cuts vector should generally be 1")
+    }  
+    if (col_quantile_by_facet == TRUE) {
       data <- data %>%
         dplyr::group_by(!!facet_var) %>%
         dplyr::mutate(!!col_var := percent_rank(!!col_var)) %>%
         dplyr::mutate(!!col_var := cut(
           !!col_var,
-          quantile_cuts,
+          col_cuts,
           right = FALSE,
           include.lowest = TRUE
         ))
       
       if (is.null(pal))
-        pal <- viridis::viridis(length(quantile_cuts) - 1)
+        pal <- viridis::viridis(length(col_cuts) - 1)
       if (is.null(legend_labels))
         labels <-
-          paste0(numeric_legend_labels(quantile_cuts * 100, 0),
+          paste0(numeric_legend_labels(col_cuts * 100, 0),
                  "\u1D57\u02B0 percentile")
       if (!is.null(legend_labels))
         labels <- legend_labels
       
     }
-    else if (quantile_by_facet == FALSE) {
-      bin_cuts <-
-        quantile(col_var_vector, probs = quantile_cuts, na.rm = TRUE)
-      if (anyDuplicated(bin_cuts) > 0) stop("quantile_cuts do not provide unique breaks")
+    else if (col_quantile_by_facet == FALSE) {
+      col_cuts <-
+        quantile(col_var_vector, probs = col_cuts, na.rm = TRUE)
+      if (anyDuplicated(col_cuts) > 0) stop("col_cuts do not provide unique breaks")
         data <-
           dplyr::mutate(data,
                         !!col_var := cut(
                           col_var_vector,
-                          bin_cuts,
+                          col_cuts,
                           right = FALSE,
                           include.lowest = TRUE
                         ))
       
-      if (is.null(pal)) pal <- viridis::viridis(length(bin_cuts) - 1)
-      if (is.null(legend_labels)) labels <- numeric_legend_labels(bin_cuts, 2)
+      if (is.null(pal)) pal <- viridis::viridis(length(col_cuts) - 1)
+      if (is.null(legend_labels)) labels <- numeric_legend_labels(col_cuts, 2)
       if (!is.null(legend_labels)) labels <- legend_labels
     }
   }