+++ R/leaflet_stars_fun.R
@@ -95,8 +95,7 @@ leaflet_stars <- function(data,
 #' @description Map of an array in leaflet that is coloured. 
 #' @param data A stars object with dimensions x and y, and 1 attribute layer with crs in wgs84 (epsg4326). Required input.
 #' @param col_method The method of colouring features, either "bin", "quantile" or "category." Defaults to "quantile". Note all numeric variables are cut to be inclusive of the min in the range, and exclusive of the max in the range (except for the final bucket which includes the highest value).
-#' @param bin_cuts A vector of bin cuts applicable where col_method of "bin" is selected. The first number in the vector should be either -Inf or 0, and the final number Inf. If NULL, 'pretty' breaks are used. Only applicable where col_method equals "bin".
-#' @param quantile_cuts A vector of probability cuts applicable where col_method of "quantile" is selected. The first number in the vector should 0 and the final number 1. Defaults to quartiles. Only applicable where col_method equals "quantile".
+#' @param col_cuts A vector of cuts to colour a numeric variable. If "bin" is selected, the first number in the vector should be either -Inf or 0, and the final number Inf. If "quantile" is selected, the first number in the vector should be 0 and the final number should be 1. Defaults to quartiles. 
 #' @param pal Character vector of hex codes, or provided objects with pal_ prefixes. Defaults to viridis.
 #' @param pal_rev Reverses the palette. Defaults to FALSE.
 #' @param opacity Sets the opacity of the grid cells. Defaults to 0.9.
@@ -109,12 +108,11 @@ leaflet_stars <- function(data,
 #' @export
 #' @examples
 #' leaflet_stars_col(example_stars_nz_no3n, 
-#'    col_method = "quantile", quantile_cuts = c(0, 0.05, 0.25, 0.5, 0.75, 0.95, 1),
+#'    col_method = "quantile", col_cuts = c(0, 0.05, 0.25, 0.5, 0.75, 0.95, 1),
 #'    title = "River modelled median nitrate-nitrogen concentrations in g/m\u00b3, 2013\u201317")
 leaflet_stars_col <- function(data,
                               col_method = "quantile",
-                              quantile_cuts = c(0, 0.25, 0.5, 0.75, 1),
-                              bin_cuts = NULL,
+                              col_cuts = NULL,
                               pal = NULL,
                               pal_rev = FALSE,
                               opacity = 1,
@@ -137,61 +135,61 @@ leaflet_stars_col <- function(data,
     no_bins <-
       max(col_var_vector, na.rm = TRUE) - min(col_var_vector, na.rm = TRUE) + 1
     max_bin_cut <- max(col_var_vector, na.rm = TRUE) + 1
-    bin_cuts <- seq(min(col_var_vector, na.rm = TRUE), max_bin_cut, 1)
+    col_cuts <- seq(min(col_var_vector, na.rm = TRUE), max_bin_cut, 1)
     if (is.null(pal))
-      pal <- pal_point_set1[1:(length(bin_cuts) - 1)]
+      pal <- pal_point_set1[1:(length(col_cuts) - 1)]
     else if (!is.null(pal))
-      pal <- pal[1:(length(bin_cuts) - 1)]
+      pal <- pal[1:(length(col_cuts) - 1)]
     if (pal_rev == TRUE)
       pal <- rev(pal)
     pal_fun <-
       colorBin(
         palette = pal,
         domain = col_var_vector,
-        bins = bin_cuts,
+        bins = col_cuts,
         right = FALSE,
         na.color = "transparent"
       )
     pal <- stringr::str_sub(pal, 1, 7)
     if (is.null(legend_labels))
-      labels <- LETTERS[1:length(bin_cuts) - 1]
+      labels <- LETTERS[1:length(col_cuts) - 1]
     else if (!is.null(legend_labels))
       labels <- legend_labels
   }
   else if (col_method == "bin") {
-    if (!is.null(bin_cuts)) {
-      if (!(dplyr::first(bin_cuts) %in% c(0, -Inf)))
+    if (!is.null(col_cuts)) {
+      if (!(dplyr::first(col_cuts) %in% c(0, -Inf)))
         warning(
-          "The first element of the bin_cuts vector should generally be 0 (or -Inf if there are negative values)"
+          "The first element of the col_cuts vector should generally be 0 (or -Inf if there are negative values)"
         )
-      if (dplyr::last(bin_cuts) != Inf)
-        warning("The last element of the bin_cuts vector should generally be Inf")
+      if (dplyr::last(col_cuts) != Inf)
+        warning("The last element of the col_cuts vector should generally be Inf")
       if (is.null(pal))
-        pal <- viridis::viridis(length(bin_cuts) - 1)
+        pal <- viridis::viridis(length(col_cuts) - 1)
       else if (!is.null(pal))
-        pal <- pal[1:(length(bin_cuts) - 1)]
+        pal <- pal[1:(length(col_cuts) - 1)]
       if (pal_rev == TRUE)
         pal <- rev(pal)
       pal_fun <-
         colorBin(
           palette = pal,
           domain = col_var_vector,
-          bins = bin_cuts,
+          bins = col_cuts,
           right = FALSE,
           na.color = "transparent"
         )
       pal <- stringr::str_sub(pal, 1, 7)
       if (is.null(legend_labels))
-        labels <- numeric_legend_labels(bin_cuts, legend_digits)
+        labels <- numeric_legend_labels(col_cuts, legend_digits)
       else if (!is.null(legend_labels))
         labels <- legend_labels
     }
-    else if (is.null(bin_cuts)) {
-      bin_cuts <- pretty(col_var_vector)
+    else if (is.null(col_cuts)) {
+      col_cuts <- pretty(col_var_vector)
       if (is.null(pal))
-        pal <- viridis::viridis(length(bin_cuts) - 1)
+        pal <- viridis::viridis(length(col_cuts) - 1)
       else if (!is.null(pal))
-        pal <- pal[1:(length(bin_cuts) - 1)]
+        pal <- pal[1:(length(col_cuts) - 1)]
       if (pal_rev == TRUE)
         pal <- rev(pal)
       pal_fun <-
@@ -204,37 +202,35 @@ leaflet_stars_col <- function(data,
         )
       pal <- stringr::str_sub(pal, 1, 7)
       if (is.null(legend_labels))
-        labels <- numeric_legend_labels(bin_cuts, legend_digits)
+        labels <- numeric_legend_labels(col_cuts, legend_digits)
       else if (!is.null(legend_labels))
         labels <- legend_labels
     }
   }
   else if (col_method == "quantile") {
-    if (dplyr::first(quantile_cuts) != 0)
-      warning("The first element of the quantile_cuts vector generally always be 0")
-    if (dplyr::last(quantile_cuts) != 1)
-      warning("The last element of the quantile_cuts vector should generally be 1")
-    if (is.null(pal))
-      pal <- viridis::viridis(length(quantile_cuts) - 1)
-    else if (!is.null(pal))
-      pal <- pal[1:(length(quantile_cuts) - 1)]
-    if (pal_rev == TRUE)
-      pal <- rev(pal)
-    bin_cuts <-
-      stats::quantile(col_var_vector, probs = quantile_cuts, na.rm = TRUE)
+    if(is.null(col_cuts)) col_cuts <- seq(0, 1, 0.25)
+    else {
+      if (dplyr::first(col_cuts) != 0) warning("The first element of the col_cuts vector generally always be 0")
+      if (dplyr::last(col_cuts) != 1) warning("The last element of the col_cuts vector should generally be 1")
+    } 
+    
+    if (is.null(pal)) pal <- viridis::viridis(length(col_cuts) - 1)
+    else if (!is.null(pal)) pal <- pal[1:(length(col_cuts) - 1)]
+    if (pal_rev == TRUE) pal <- rev(pal)
+    col_cuts <- stats::quantile(col_var_vector, probs = col_cuts, na.rm = TRUE)
+    
     pal_fun <-
       colorBin(
         palette = pal,
         domain = col_var_vector,
-        bins = bin_cuts,
+        bins = col_cuts,
         right = FALSE,
         na.color = "transparent"
       )
+    
     pal <- stringr::str_sub(pal, 1, 7)
-    if (is.null(legend_labels))
-      labels <- numeric_legend_labels(bin_cuts, legend_digits)
-    else if (!is.null(legend_labels))
-      labels <- legend_labels
+    if (is.null(legend_labels)) labels <- numeric_legend_labels(col_cuts, legend_digits)
+    else if (!is.null(legend_labels)) labels <- legend_labels
   }
   
   data <- methods::as(data, "Raster")